requirements.txt
Flask
pycryptodome
soundfile
scipy
numpy